+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- [#1752](https://github.com/teambit/bit/issues/1752) fix dependency links generation when originally there were multiple link files
+
 ## [14.1.4-dev.5] - 2019-06-25
 
 - add experimental `post-add` hook