+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- add a custom entry point file for Angular components
+- [#1750](https://github.com/teambit/bit/issues/1750) improve the output to clarify when a dependency package is missing
+- [#1752](https://github.com/teambit/bit/issues/1752) fix dependency links generation when originally there were multiple link files
+
+## [14.1.4-dev.5] - 2019-06-25
+
+- add experimental `post-add` hook
+
 ## [14.1.4-dev.4] - 2019-06-24
 
 - add `--no-cache` flag to `bit ci-update` command